requirements.txt
asgiref==3.8.1
Django==5.2
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
phonenumbers==9.0.3
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
